Rework routing and add user auth based fetching for routes

+++ controllers/capture.go
@@ -5,14 +5,37 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/lithammer/shortuuid/v4"
+	"github.com/luminarapp/server/auth"
 	"github.com/luminarapp/server/models"
 )
+
+// GET /captures
+func GetUserCaptures(c *gin.Context) {
+	user, err := auth.ExtractTokenID(c)
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Get captures
+	captures, err := models.GetCapturesByUserId(user)
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// TODO: Paginate captures
+	c.JSON(http.StatusOK, gin.H{"data": captures})
+}
 
 // GET /captures/:id
 func GetCapture(c *gin.Context) {
 	var capture models.Capture
 
-	if err := models.DB.Where("id = ?", c.Param("id")).First(&capture).Error; err != nil {
+	// Get capture with preloaded comments
+	if err := models.DB.Preload("Comments").Where("id = ?", c.Param("id")).First(&capture).Error; err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "capture id not found"})
 		return
 	}
@@ -29,10 +52,28 @@ func CreateCapture(c *gin.Context) {
 		return
 	}
 
+	// Get user
+	userId, err := auth.ExtractTokenID(c)
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Check if collection exists
+	var collection models.Collection
+
+	if err := models.DB.Where("id = ?", payload.CollectionID).First(&collection).Error; err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "collection id not found"})
+		return
+	}
+
+	// Create capture
 	capture := models.Capture{
 		ID: shortuuid.New(),
-		Author: payload.Author,
-		Source: payload.Source,
+		UserID: userId,
+		CollectionID: payload.CollectionID,
+		Reference: payload.Reference,
 	}
 	
 	if err := models.DB.Create(&capture).Error; err != nil {
@@ -52,6 +93,20 @@ func DeleteCapture(c *gin.Context) {
 		return
 	}
 
+	// Authenticate user
+	userId, err := auth.ExtractTokenID(c)
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if capture.UserID != userId {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "use not allowed to delete capture"})
+		return
+	}
+
+	// Delete capture
 	if err := models.DB.Delete(&capture).Error; err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
@@ -59,3 +114,6 @@ func DeleteCapture(c *gin.Context) {
 
 	c.JSON(http.StatusOK, gin.H{"data": capture})
 }
+
+// PATCH /captures/:id
+// TODO
